requirements.txt
requests==2.28.1
pyyaml==6.0